fix : User byte->long , Swagger 문구 수정

+++ src/main/java/com/devhive03/Model/DTO/User/UserDTO.java
@@ -14,9 +14,9 @@ public class UserDTO {
     private String phoneNumber;
     private String profilePhoto;
     private String introduction;
-    private Byte membership;
-    private Byte certification;
-    private Byte ratingState;
+    private Long membership;
+    private Long certification;
+    private Long ratingState;
 
     public static UserDTO of(User user) {
         return UserDTO.builder()
